GVT-2477: Review fixes
