+++ ui/src/tool-bar/tool-bar-container.tsx
@@ -1,12 +1,11 @@
 import * as React from 'react';
 import { ToolBar } from 'tool-bar/tool-bar';
-import { useCommonDataAppSelector, useTrackLayoutAppSelector } from 'store/hooks';
+import { useTrackLayoutAppSelector } from 'store/hooks';
 import { createDelegates } from 'store/store-utils';
 import { trackLayoutActionCreators } from 'track-layout/track-layout-slice';
 
 export const ToolBarContainer: React.FC = () => {
     const trackLayoutState = useTrackLayoutAppSelector((state) => state);
-    const changeTimes = useCommonDataAppSelector((state) => state.changeTimes);
     const delegates = React.useMemo(() => createDelegates(trackLayoutActionCreators), []);
 
     return (
@@ -17,7 +16,6 @@ export const ToolBarContainer: React.FC = () => {
             onUnselect={delegates.onUnselect}
             onPublishTypeChange={delegates.onPublishTypeChange}
             onOpenPreview={() => delegates.onLayoutModeChange('PREVIEW')}
-            changeTimes={changeTimes}
             onStopLinking={delegates.stopLinking}
             linkingState={trackLayoutState.linkingState}
             splittingState={trackLayoutState.splittingState}
